refactor: merge extraction and town leftover test-support expectations (#4410)
Slice B relocates town/extractor run bodies into consumer tests and folds pin types into siblings that stay under the 220-line cap, deleting the four leftover expectations modules.
