+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -29,7 +29,6 @@ export 'src/non_gp_auto_offers_test_support.dart';
 export 'src/non_gp_extraction_pins.dart';
 export 'src/non_gp_extraction_scenarios.dart';
 export 'src/extraction_fixture_support.dart';
-export 'src/resource_extractor_expectations.dart';
 export 'src/resource_extractor_scenarios.dart';
 export 'src/scenario_runner.dart';
 export 'src/sea_transport_scenarios.dart';
@@ -38,12 +37,11 @@ export 'src/tile_extraction_pipeline_scenarios.dart';
 export 'src/trade_cargo_capacity_scenarios.dart';
 export 'src/treasury_scenarios/treasury_scenarios.dart';
 export 'src/projected_cost_engine_scenarios.dart';
-export 'src/province_extraction_snapshot_expectations.dart';
+export 'src/province_extraction_snapshot_pins.dart';
 export 'src/province_extraction_snapshot_scenarios.dart';
 export 'src/purchased_tile_fixture_support.dart';
 export 'src/purchased_tile_scenarios/purchased_tile_expectations.dart';
 export 'src/purchased_tile_scenarios/purchased_tile_scenarios.dart';
-export 'src/town_manufacturing_bonus_expectations.dart';
 export 'src/town_manufacturing_bonus_scenarios.dart';
 export 'src/trade_interception_scenarios.dart';
 export 'src/worker_action_cost_scenarios.dart';
