+++ packages/colonizethis_economy_test_support/lib/src/province_extraction_snapshot_pins.dart
@@ -1,5 +1,5 @@
 // dart format off
-// Compact province Extraction/Available assertions (Refs #4002, #4014).
+// Compact province Extraction/Available assertions (Refs #4002, #4014, #4410).
 
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_economy/colonizethis_economy.dart';
